More improvements to the test runner,  * an autoloader is provided for locating classes    * checks if it has a class hash 	   * loads a cached copy if present, if the class is defined it requires it and returns      * parses the agavi and, if present, webapp dirs for classes 		 * stores the hash in a static var 		 * if the cache var points to a writable file, it will write out the hash to it (you probably need to make the tests dir writable) 	 * requires the class
This should be a nice fall back, I think. You can include the files you need in your test,
but if you dont we should still be able to find them lickity split. Generating mocks of 
classes that've not yet been included doesnt work without 'fixing' simpletest to not 
suppress autoloading when checking for the class, I've submitted a request for this to be
fixed but you may need to make sure your classes are loaded before attempting to mock them, 
in the mean time.