make depend tokens available via validation report
The tokens provided by validators for other validators to depend on are
not accessible later on in the request. This changeset makes the depend
tokens available via the validation report and the validation report
query. This change should be backwards compatible and does not change how
the validation works atm.