Remove extra quotations from batch file
The command

SET AGAVI_SOURCE_DIRECTORY="@PEAR-DIR@/agavi"

actually includes the quotations in the value. This is not desired, and
it resulted in an error when I attempted to run agavi for the first
time.

The error was:

'""D:\Programs\PHP' is not recognized as an internal or external
command, operable program or batch file.