prevent hhvm from running the tests in isolation
It is not important to run the test the isolation on hhvm,
since the isolation check tests for dependency problems in the tests
and not any real functionality. And since running the tests that
way on hhvm is extremely slow, we need to disable them to be able
to run the tests on travis at all.